+++ Sources/PaymentsCore/api/EligibilityAPI.swift
@@ -1,12 +1,15 @@
 import Foundation
 
-public class EligibilityAPI {
+/// API that return merchant's eligibility for different payment methods: Venmo, PayPal, PayPal Credit, Pay Later & credit card
+class EligibilityAPI {
 
     private let graphQLClient: GraphQLClient
     private let apiClient: APIClient
     private let coreConfig: CoreConfig
 
-    public convenience init(coreConfig: CoreConfig) {
+    /// Initialize the eligibility API to check for payment methods eligibility
+    /// - Parameter coreConfig: configuration object
+    convenience init(coreConfig: CoreConfig) {
         self.init(
             coreConfig: coreConfig,
             apiClient: APIClient(coreConfig: coreConfig),
@@ -20,7 +23,9 @@ public class EligibilityAPI {
         self.graphQLClient = graphQLClient
     }
 
-    public func checkEligibility() async throws -> Result<Eligibility, Error> {
+    /// Checks merchants eligibility for different payment methods.
+    /// - Returns: a result object with either eligibility or an error
+    func checkEligibility() async throws -> Result<Eligibility, Error> {
         let clientID = try await apiClient.getClientID()
         let fundingEligibilityQuery = FundingEligibilityQuery(
             clientID: clientID,