+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Unreleased Version
 * PayPalNativeCheckout
     * Update to version 0.109.0 that fixes shipping callback bug
+* Bump minimum iOS deployment to 14+ and swift to 5.7+
 
 ## 0.0.1 (2022-07-21)