+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 ## Unreleased Version
 * PayPalNativeCheckout
     * Update to version 0.109.0 that fixes shipping callback bug
-* Bump minimum iOS deployment to 14+ and swift to 5.7+
+* Bump minimum supported deployment target to iOS 14+
+* Require Xcode 14
+    * Bump Swift Tools Version to 5.7 for CocoaPods & SPM
 
 ## 0.0.1 (2022-07-21)