+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 
 # PayPal iOS SDK Release Notes
 
+## unreleased
+* PayPalNativeCheckout
+    * Update to version 0.109.0 that fixes shipping callback bug
+* Bump minimum supported deployment target to iOS 14+
+* Require Xcode 14
+    * Bump Swift Tools Version to 5.7 for CocoaPods & SPM
+
 ## 0.0.1 (2022-07-21)
 
 * Card